+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Add support for automatic AWS IAM authentication using EC2 instance metadata. CNJR-10957
+- Add Static Secrets V2 API. CNJR-10824
 
 ## [0.13.4] - 2025-09-08